+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Feature) JobScheduler Coverage
 - (Feature) JobScheduler Volumes, Probes, Lifecycle and Ports integration
 - (Feature) Merge ArangoDB Usage Metrics
+- (Bugfix) Check Connection to the ArangoDB before creating Backup
 
 ## [1.2.38](https://github.com/arangodb/kube-arangodb/tree/1.2.38) (2024-02-22)
 - (Feature) Extract GRPC Server